requirements.txt
fastapi
pydantic
uvicorn
poe-api